@@ -69,24 +69,28 @@
 
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 		<github.global.server>github</github.global.server>
 		<project.build.timestamp>${maven.build.timestamp}</project.build.timestamp>
 
 		<!-- Dependency Versions -->
 
-		<did-common-java.version>0.7.0</did-common-java.version>
+		<did-common-java.version>1.4.0</did-common-java.version>
 
-		<eclipse.jetty.version>9.4.44.v20210927</eclipse.jetty.version>
-		<slf4j.version>1.7.25</slf4j.version>
-		<apache.log4j.version>2.17.1</apache.log4j.version>
-		<fasterxml.jackson.version>2.11.2</fasterxml.jackson.version>
-		<gson.version>2.8.6</gson.version>
-		<commons-codec.version>1.14</commons-codec.version>
-		<httpclient.version>4.5.13</httpclient.version>
-		<javax.servlet.version>2.5</javax.servlet.version>
-		<springframework.version>5.3.14</springframework.version>
-		<junit.version>4.13.1</junit.version>
+		<slf4j.version>2.0.6</slf4j.version>
+		<gson.version>2.10.1</gson.version>
+		<jackson.databind.version>2.14.2</jackson.databind.version>
+		<jackson.core.version>2.14.2</jackson.core.version>
+		<jackson.annotations.version>2.14.2</jackson.annotations.version>
+		<commons-codec.version>1.15</commons-codec.version>
+		<httpclient.version>4.5.14</httpclient.version>
+
+		<springframework.version>6.0.0</springframework.version>
+		<jakarta.servlet-api.version>5.0.0</jakarta.servlet-api.version>
+
+		<junit-jupiter.version>5.9.2</junit-jupiter.version>
+		<mockito-core.version>3.4.4</mockito-core.version>
+		<junit.version>4.13.2</junit.version>
 
 		<!-- Plugin Versions -->
 
