@@ -8,12 +8,12 @@
 	<parent>
 		<groupId>decentralized-identity</groupId>
 		<artifactId>uni-resolver</artifactId>
-		<version>0.6-SNAPSHOT</version>
+		<version>0.13-SNAPSHOT</version>
 	</parent>
 
     <properties>
-        <java.version>11</java.version>
-        <spring.boot.version>2.6.2</spring.boot.version>
+        <java.version>17</java.version>
+        <spring.boot.version>3.0.0</spring.boot.version>
     </properties>
 
 	<build>
