@@ -3,7 +3,7 @@ package uniresolver.result;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSetter;
-import foundation.identity.did.jsonld.DIDContexts;
+import org.apache.http.entity.ContentType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uniresolver.ResolutionException;
@@ -18,6 +18,7 @@ public abstract class ResolveResult implements Result {
     public static final URI DEFAULT_JSONLD_CONTEXT = URI.create("https://w3id.org/did-resolution/v1");
 
     public static final String MEDIA_TYPE = "application/ld+json;profile=\"https://w3id.org/did-resolution\"";
+    public static final ContentType CONTENT_TYPE = ContentType.parse(MEDIA_TYPE);
 
     private static final Logger log = LoggerFactory.getLogger(ResolveResult.class);
 
@@ -52,14 +53,26 @@ public abstract class ResolveResult implements Result {
     }
 
     /*
+     * Media type methods
+     */
+
+    public static boolean isResolveResultMediaType(ContentType mediaType) {
+        boolean isResolveResultMimeTypeEquals = CONTENT_TYPE.getMimeType().equals(mediaType.getMimeType());
+        boolean isResolveResultProfileEquals = CONTENT_TYPE.getParameter("profile").equals(mediaType.getParameter("profile"));
+        return isResolveResultMimeTypeEquals && isResolveResultProfileEquals;
+    }
+
+    /*
      * Conversion
      */
 
     public ResolveDataModelResult toResolveDataModelResult() throws ResolutionException {
         if (this.resolveDataModelResult == null) {
             if (this instanceof ResolveDataModelResult) {
+                if (log.isDebugEnabled()) log.debug("toResolveDataModelResult(): this=ResolveDataModelResult");
                 this.resolveDataModelResult = (ResolveDataModelResult) this;
             } else if (this instanceof ResolveRepresentationResult) {
+                if (log.isDebugEnabled()) log.debug("toResolveDataModelResult(): this=ResolveRepresentationResult");
                 ResolveDataModelResult resolveDataModelResult = Conversion.convertToResolveDataModelResult((ResolveRepresentationResult) this);
                 this.resolveDataModelResult = resolveDataModelResult;
             }
@@ -71,6 +84,7 @@ public abstract class ResolveResult implements Result {
         if (representationMediaType == null) throw new NullPointerException();
         if (! this.resolveRepresentationResults.containsKey(representationMediaType)) {
             if (this instanceof ResolveRepresentationResult) {
+                if (log.isDebugEnabled()) log.debug("toResolveRepresentationResult(): this=ResolveRepresentationResult, representationMediaType=" + representationMediaType + ", this.contentType=" + ((ResolveRepresentationResult) this).getContentType());
                 if (representationMediaType.equals(((ResolveRepresentationResult) this).getContentType())) {
                     this.resolveRepresentationResults.put(representationMediaType, (ResolveRepresentationResult) this);
                 } else {
@@ -81,6 +95,7 @@ public abstract class ResolveResult implements Result {
                     this.resolveRepresentationResults.put(representationMediaType, resolveRepresentationResult);
                 }
             } else if (this instanceof ResolveDataModelResult) {
+                if (log.isDebugEnabled()) log.debug("toResolveRepresentationResult(): this=ResolveDataModelResult");
                 ResolveRepresentationResult resolveRepresentationResult = Conversion.convertToResolveRepresentationResult((ResolveDataModelResult) this, representationMediaType);
                 this.resolveRepresentationResults.put(representationMediaType, resolveRepresentationResult);
             }
