@@ -1,43 +1,49 @@
 package uniresolver.web.servlet;
 
+import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
+import java.util.Map;
 
 public class ServletUtil {
 
 	private static final Logger log = LoggerFactory.getLogger(ServletUtil.class);
 
-	static void sendResponse(HttpServletResponse response, int status, String contentType, Object body) throws IOException {
+	public static void sendResponse(HttpServletResponse response, int status, Map<String, String> headers, byte[] body) throws IOException {
 
-		if (log.isDebugEnabled()) log.debug("Sending response with status " + status + " and content type " + contentType + " and body (" + (body == null ? null : body.getClass()) + ")");
+		if (log.isInfoEnabled()) log.info("Sending response with status " + status + " and headers " + headers + " and body (" + (body == null ? null : body.length) + ")");
 
 		response.setStatus(status);
-		if (contentType != null) response.setContentType(contentType);
+		if (headers != null) for (Map.Entry<String, String> header : headers.entrySet()) response.setHeader(header.getKey(), header.getValue());
 		response.setHeader("Access-Control-Allow-Origin", "*");
 
-		if (body instanceof String) {
-
-			PrintWriter printWriter = response.getWriter();
-			printWriter.write((String) body);
-			printWriter.flush();
-			printWriter.close();
-		} else if (body instanceof byte[]) {
-
+		if (body != null) {
 			OutputStream outputStream = response.getOutputStream();
-			outputStream.write((byte[]) body);
+			outputStream.write(body);
 			outputStream.flush();
 			outputStream.close();
-		} else if (body != null) {
-
-			PrintWriter printWriter = response.getWriter();
-			printWriter.write(body.toString());
-			printWriter.flush();
-			printWriter.close();
 		}
+
+		response.flushBuffer();
+	}
+
+	public static void sendResponse(HttpServletResponse response, int status, String contentType, byte[] body) throws IOException {
+		sendResponse(response, status, Map.of("Content-Type", contentType), body);
+	}
+
+	public static void sendResponse(HttpServletResponse response, int status, String contentType, String body) throws IOException {
+		sendResponse(response, status, Map.of("Content-Type", contentType), body == null ? null : body.getBytes(StandardCharsets.UTF_8));
+	}
+
+	public static void sendResponse(HttpServletResponse response, int status, byte[] body) throws IOException {
+		sendResponse(response, status, (Map<String, String>) null, body);
+	}
+
+	public static void sendResponse(HttpServletResponse response, int status, String body) throws IOException {
+		sendResponse(response, status, (Map<String, String>) null, body == null ? null : body.getBytes(StandardCharsets.UTF_8));
 	}
 }
